2013-02-01  Jeremy Bennett <jeremy.bennett@embecosm.com>
	* build-all.sh: Add option to control building of unified source
	tree.

Conflicts:
	build-all.sh (resolved).

+++ build-all.sh
@@ -29,6 +29,7 @@
 #                  [--symlink-dir <symlink_dir>]
 #                  [--auto-pull | --no-auto-pull]
 #                  [--auto-checkout | --no-auto-checkout]
+#                  [--unisrc | --no-unisrc]
 #                  [--elf32 | --no-elf32] [--linux | --no-linux]
 #                  [--datestamp-install]
 #                  [--comment-install <comment>] [--big-endian]
@@ -105,6 +106,11 @@
 
 #     If specified, a "git pull" will be done in each component repository
 #     after checkout to ensure the latest code is in use. Default is to pull.
+
+# --unisrc | --no-unisrc
+
+#     If --unisrc is specified, rebuild the unified source tree. If
+#     --no-unisrc is specified, do not rebuild it. The default is --unisrc.
 
 # --elf32 | --no-elf32
 
@@ -207,6 +213,7 @@ build_pathnm ()
 # Set defaults for some options
 autocheckout="--auto-checkout"
 autopull="--auto-pull"
+do_unisrc="--unisrc"
 elf32="--elf32"
 linux="--linux"
 
@@ -252,6 +259,10 @@ case ${opt} in
 
     --auto-pull | --no-auto-pull)
 	autopull=$1
+	;;
+
+    --unisrc | --no-unisrc)
+	do_unisrc=$1
 	;;
 
     --elf32 | --no-elf32)
@@ -303,6 +314,7 @@ case ${opt} in
 	echo "                      [--symlink-dir <symlink_dir>]"
 	echo "                      [--auto-checkout | --no-auto-checkout]"
         echo "                      [--auto-pull | --no-auto-pull]"
+        echo "                      [--unisrc | --no-unisrc]"
         echo "                      [--elf32 | --no-elf32]"
         echo "                      [--linux | --no-linux]"
 	echo "                      [--datestamp-install]"
@@ -388,7 +400,7 @@ fi
 PARALLEL="-j ${jobs} -l ${load}"
 
 # All the things we export to the scripts
-export UNISRC=unisrc-4.8
+export UNISRC=unisrc
 export ARC_GNU
 export LINUXDIR
 export INSTALLDIR
@@ -400,8 +412,7 @@ export PARALLEL
 cd ${builddir}
 
 # Set up a logfile
-mkdir -p ${PWD}/logs
-logfile="$(echo "${PWD}")/logs/all-build-$(date -u +%F-%H%M).log"
+logfile="$(echo "${PWD}")/all-build-$(date -u +%F-%H%M).log"
 rm -f "${logfile}"
 
 # Checkout the correct branch for each tool
@@ -410,7 +421,7 @@ echo "======================" >> "${logfile}"
 
 echo "Checking out GIT trees ..."
 if ! ${ARC_GNU}/toolchain/arc-versions.sh ${autocheckout} ${autopull} \
-      >> ${logfile} 2>&1
+    >> ${logfile} 2>&1
 then
     echo "ERROR: Failed to checkout GIT versions of tools"
     exit 1
@@ -418,25 +429,22 @@ fi
 
 # Make a unified source tree in the build directory. Note that later versions
 # override earlier versions with the current symlinking version.
-echo "Linking unified tree" >> "${logfile}"
-echo "====================" >> "${logfile}"
-
-echo "Linking unified tree ..."
-component_dirs="gdb newlib gcc binutils"
-# component_dirs="gdb newlib binutils gcc"
-rm -rf ${UNISRC}
-
-if ! mkdir -p ${UNISRC}
+if [ "x${do_unisrc}" = "x--unisrc" ]
 then
-    echo "ERROR: Failed to create ${UNISRC}"
-    exit 1
-fi
+    echo "Linking unified tree" >> "${logfile}"
+    echo "====================" >> "${logfile}"
 
-if ! ${ARC_GNU}/toolchain/symlink-all.sh ${UNISRC} \
-    "${component_dirs}" >> "${logfile}" 2>&1
-then
-    echo "ERROR: Failed to symlink ${UNISRC}"
-    exit 1
+    echo "Linking unified tree ..."
+    component_dirs="gdb newlib gcc binutils"
+    # component_dirs="gdb newlib binutils gcc"
+     rm -rf ${UNISRC}
+
+    if ! ${ARC_GNU}/toolchain/symlink-all.sh ${UNISRC} \
+        "${component_dirs}" >> "${logfile}" 2>&1
+    then
+        echo "ERROR: Failed to symlink ${UNISRC}"
+        exit 1
+    fi
 fi
 
 # Optionally build the arc-elf32- tool chain
